requirements.txt
atlassian-python-api
jmespath
markdownify
pandas
pydantic-settings
pyyaml
questionary
rich
tabulate
tqdm
typer